+++ sandbox/tests/camera/lerp.ts
@@ -11,20 +11,20 @@ game.add(actor);
 
 game.start().then(() => {});
 
-var easingFn = ex.EasingFunctions.EaseInOutQuad;
+var easingFn = ex.easeInOutQuad;
 
 game.input.pointers.primary.on('down', (evt: ex.PointerEvent) => {
   game.currentScene.camera.move(new ex.Vector(evt.worldPos.x, evt.worldPos.y), 500, easingFn).then((v) => onLerpEnd(v));
   document.getElementById('lerp-false').style.display = 'none';
   document.getElementById('lerp-true').style.display = 'inline';
 });
-document.getElementById('move-ease-in-out-quad').addEventListener('click', moveCameraEase.bind(this, ex.EasingFunctions.EaseInOutQuad));
-document.getElementById('move-ease-in-quad').addEventListener('click', moveCameraEase.bind(this, ex.EasingFunctions.EaseInQuad));
-document.getElementById('move-ease-out-quad').addEventListener('click', moveCameraEase.bind(this, ex.EasingFunctions.EaseOutQuad));
-document.getElementById('move-ease-in-out-cubic').addEventListener('click', moveCameraEase.bind(this, ex.EasingFunctions.EaseInOutCubic));
-document.getElementById('move-ease-in-cubic').addEventListener('click', moveCameraEase.bind(this, ex.EasingFunctions.EaseInCubic));
-document.getElementById('move-ease-out-cubic').addEventListener('click', moveCameraEase.bind(this, ex.EasingFunctions.EaseOutCubic));
-document.getElementById('move-ease-linear').addEventListener('click', moveCameraEase.bind(this, ex.EasingFunctions.Linear));
+document.getElementById('move-ease-in-out-quad').addEventListener('click', moveCameraEase.bind(this, ex.easeInOutQuad));
+document.getElementById('move-ease-in-quad').addEventListener('click', moveCameraEase.bind(this, ex.easeInQuad));
+document.getElementById('move-ease-out-quad').addEventListener('click', moveCameraEase.bind(this, ex.easeOutQuad));
+document.getElementById('move-ease-in-out-cubic').addEventListener('click', moveCameraEase.bind(this, ex.easeInOutCubic));
+document.getElementById('move-ease-in-cubic').addEventListener('click', moveCameraEase.bind(this, ex.easeInCubic));
+document.getElementById('move-ease-out-cubic').addEventListener('click', moveCameraEase.bind(this, ex.easeOutCubic));
+document.getElementById('move-ease-linear').addEventListener('click', moveCameraEase.bind(this, ex.linear));
 document.getElementById('move-xy').addEventListener('click', moveCameraViaXY);
 
 var sw = true;