+++ sandbox/tests/tilemap/tilemap.ts
@@ -57,9 +57,9 @@ game.input.pointers.primary.on('down', (evt: ex.PointerEvent) => {
 });
 
 game.currentScene.onActivate = async () => {
-  await game.currentScene.camera.move(ex.Vector.Zero.clone(), 2000, ex.EasingFunctions.EaseInOutCubic);
+  await game.currentScene.camera.move(ex.Vector.Zero.clone(), 2000, ex.easeInOutCubic);
   console.log(tm.getOnScreenTiles());
-  await game.currentScene.camera.move(new ex.Vector(800, 600), 4000, ex.EasingFunctions.EaseInOutCubic);
+  await game.currentScene.camera.move(new ex.Vector(800, 600), 4000, ex.easeInOutCubic);
   console.log(tm.getOnScreenTiles());
 
   await ex.coroutine(
@@ -79,7 +79,7 @@ game.currentScene.onActivate = async () => {
   console.log(tm.getOnScreenTiles());
   await game.currentScene.camera.zoomOverTime(1, 1000);
   console.log(tm.getOnScreenTiles());
-  await game.currentScene.camera.move(tm.pos, 2000, ex.EasingFunctions.EaseInOutCubic);
+  await game.currentScene.camera.move(tm.pos, 2000, ex.easeInOutCubic);
   console.log(tm.getOnScreenTiles());
   await game.currentScene.camera.zoomOverTime(2, 1000);
   console.log(tm.getOnScreenTiles());